Make hardening tests DB-agnostic
Use schema helpers and explicit foreign key disabling so the dedupe and installation hardening tests work across SQLite, MySQL, and PostgreSQL. The installation test case now honors the active DB_* connection, and race-related test imports were updated accordingly.

+++ tests/Feature/PayPalCaptureSemanticsTest.php
@@ -237,7 +237,7 @@ it('preserves an earlier partial capture when a follow-up capture fails', functi
 
     expect($payment->fresh()->status)->toBe(PaymentStatus::PartiallyCaptured)
         ->and($payment->fresh()->captured_minor)->toBe(4000)
-        ->and(PaymentTransaction::query()
+        ->and((int) PaymentTransaction::query()
             ->where('payment_id', $payment->id)
             ->where('type', PaymentTransactionType::Capture)
             ->where('status', 'succeeded')
